+++ java/openinference-semantic-conventions/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.1.11](https://github.com/Arize-ai/openinference/compare/java-openinference-semantic-conventions-v0.1.10...java-openinference-semantic-conventions-v0.1.11) (2026-03-23)
+
+
+### Features
+
+* Add missing Java conventions + fix dependencies ([#2870](https://github.com/Arize-ai/openinference/issues/2870)) ([6ff2d19](https://github.com/Arize-ai/openinference/commit/6ff2d19da9cdcaf2b46155af679819ddbf991832))
+
 ## [0.1.10](https://github.com/Arize-ai/openinference/compare/java-openinference-semantic-conventions-v0.1.9...java-openinference-semantic-conventions-v0.1.10) (2026-03-11)